requirements.txt
langgraph>=0.2.0
langchain-aws>=0.2.0
langchain-core>=0.3.0
boto3>=1.35.0
streamlit>=1.40.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Dev dependencies
ruff>=0.8.0
pytest>=8.0.0
mypy>=1.13.0
